requirements.txt
dspy_ai==2.4.9
wikipedia==1.4.0
sentence_transformers
toml
langchain-text-splitters
trafilatura
langchain-huggingface
qdrant-client
langchain-qdrant
numpy==1.26.4
streamlit==1.31.1
streamlit-card
markdown
unidecode
extra-streamlit-components==0.1.60
streamlit_extras
deprecation==2.1.0
st-pages==0.4.5
streamlit-float
streamlit-option-menu
watchdog